% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_vectors.R
\name{id_vector}
\alias{id_vector}
\title{id vector with sequence of integers}
\usage{
id_vector(size, start = 1)
}
\arguments{
\item{size}{integer, size of the output vector}

\item{start}{integer, value of the first element}
}
\description{
id vector with sequence of integers
}
\examples{
id_vector(10, 2)
}
