% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_types.R
\name{integer}
\alias{integer}
\title{Generate random integer}
\usage{
integer(min = 0, max = 999999)
}
\arguments{
\item{min}{integer minimum}

\item{max}{integer maximum}
}
\value{
random integer
}
\description{
Generate random integer
}
\examples{
integer(min = 2, max = 10)
}
