% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data_generator.R
\name{MockDataGenerator}
\alias{MockDataGenerator}
\title{MockDataGenerator R6 Class}
\description{
MockDataGenerator R6 Class

MockDataGenerator R6 Class
}
\details{
Object that stores mock data configurations and generated datasets
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MockDataGenerator$new()}}
\item \href{#method-generate_data}{\code{MockDataGenerator$generate_data()}}
\item \href{#method-get_data}{\code{MockDataGenerator$get_data()}}
\item \href{#method-clone}{\code{MockDataGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new MockDataGenerator object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockDataGenerator$new(configuration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{configuration}}{list; list with datasets configurations.
Check \code{random_data_frame} for configuration details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new MockDataGenerator object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_data"></a>}}
\if{latex}{\out{\hypertarget{method-generate_data}{}}}
\subsection{Method \code{generate_data()}}{
Generate and save a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockDataGenerator$generate_data(data_name, size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{string; name of data set to generate}

\item{\code{size}}{integer; number of rows in generated data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get a dataset (if does not exist, generate it)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockDataGenerator$get_data(data_name, size = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{string; data set name to retrieve}

\item{\code{size}}{integer; size of dataset (if not generated before)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
mock dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockDataGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
