% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{random_data_frame}
\alias{random_data_frame}
\title{Generate a random data frame from given configuration}
\usage{
random_data_frame(configuration)
}
\arguments{
\item{configuration}{list; a configuration of columns with data frame size
passed as "size" and all arguments required by vector generator passed as
sublists of sublist "columns"}
}
\value{
data.frame
}
\description{
Generate a random data frame from given configuration
}
\examples{
conf <- list(
 size = 10,
 columns = list(
   first_column = list(
     type = "string",
     length = 3
   ),
   second_column = list(
     type = "integer",
     max = 10
   )
 )
)

random_data_frame(conf)
}
