% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{random_date_vector}
\alias{random_date_vector}
\title{Get random date vector from an interval}
\usage{
random_date_vector(size, min_date, max_date, format = NULL, unique = FALSE)
}
\arguments{
\item{size}{integer, vector length}

\item{min_date}{character or date, beginning of the time interval to sample from}

\item{max_date}{character or date, ending of the time interval to sample from}

\item{format}{character, check \code{\link[base]{strptime}} for details}

\item{unique}{boolean, should the output be unique?}
}
\description{
Get random date vector from an interval
}
\examples{
random_date_vector(12, "2012-12-04", "2020-10-31")
}
