% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_types.R
\name{random_from_set}
\alias{random_from_set}
\title{Choose random element from set}
\usage{
random_from_set(set)
}
\arguments{
\item{set}{vector, set of values to choose from}
}
\value{
a single element from a given set
}
\description{
Choose random element from set
}
\examples{
random_from_set(c("a", "b", "c"))
}
