% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_types.R
\name{random_integer}
\alias{random_integer}
\title{Generate random integer}
\usage{
random_integer(min = 0, max = 999999)
}
\arguments{
\item{min}{integer, minimum}

\item{max}{integer, maximum}
}
\value{
random integer
}
\description{
Generate random integer
}
\examples{
random_integer(min = 2, max = 10)
}
