% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_types.R
\name{random_numeric}
\alias{random_numeric}
\title{Generate random numeric}
\usage{
random_numeric(min = 0, max = 999999)
}
\arguments{
\item{min}{numeric, minimum}

\item{max}{numeric, maximum}
}
\value{
random numeric
}
\description{
Generate random numeric
}
\examples{
random_numeric(min = 1.5, max = 4.45)
}
