% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_types.R
\name{random_string}
\alias{random_string}
\title{Generate random string}
\usage{
random_string(length = NULL, pattern = "[A-Za-z0-9]")
}
\arguments{
\item{length}{integer or NULL (default) string length.
If NULL, length will be random}

\item{pattern}{string pattern for string to follow}
}
\value{
random string
}
\description{
Generate random string
}
\examples{
random_string(length = 5)
}
