% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{random_time_vector}
\alias{random_time_vector}
\title{Get random time vector from an interval}
\usage{
random_time_vector(
  size,
  min_time = "00:00:00",
  max_time = "23:59:59",
  resolution = "seconds",
  unique = FALSE
)
}
\arguments{
\item{size}{integer, vector length}

\item{min_time}{character, beginning of the time interval to sample from}

\item{max_time}{character, ending of the time interval to sample from}

\item{resolution}{character, one of "seconds", "minutes", "hours", time resolution}

\item{unique}{boolean, should the output be unique?}
}
\description{
Get random time vector from an interval
}
\examples{
random_time_vector(12, "12:23:00", "15:48:32")
}
