% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{random_vector}
\alias{random_vector}
\title{Generate random vector of desired type}
\usage{
random_vector(size, type, ...)
}
\arguments{
\item{size}{integer vector length}

\item{type}{"integer", "string", "boolean" or "numeric" type of vector values}

\item{...}{arguments passed to function responsible for generating values.
Check \code{random_integer}, \code{random_string}, \code{random_boolean} and
\code{random_numeric} for details}
}
\value{
vector of random values of chosen type
}
\description{
Generate random vector of desired type
}
\examples{
random_vector(5, "boolean")
random_vector(10, "numeric", min = 1.5, max = 5)
random_vector(4, "string", length = 4, pattern = "[ACGT]")
random_vector(2, "integer", max = 10)
}
