% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{special_vector}
\alias{special_vector}
\title{Wrapper that allows generating a special type vector}
\usage{
special_vector(size, type, ...)
}
\arguments{
\item{size}{integer, vector length}

\item{type}{type of vector, one of: "id"}

\item{...}{additional arguments for vector function}
}
\description{
Wrapper that allows generating a special type vector
}
\examples{
special_vector(10, "id", start = 3)
}
