% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_types.R
\name{string}
\alias{string}
\title{Generate random string}
\usage{
string(length, pattern = "[A-Za-z0-9]")
}
\arguments{
\item{length}{integer string length}

\item{pattern}{string pattern for string to follow}
}
\description{
Generate random string
}
