% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_vectors.R
\name{distribution_vector}
\alias{distribution_vector}
\title{vector of values that follow specified distribution}
\usage{
distribution_vector(size, distribution_type, distribution_arguments = list())
}
\arguments{
\item{size}{integer, size of the output vector}

\item{distribution_type}{character, type of distribution. You can use
direct function name, e.g. "rnorm" or a regular name (e.g. "normal", "gaussian").
All standard distributions from {stats} package are covered. For a list
check \link[stats]{Distributions}}

\item{distribution_arguments}{list of arguments required by the distribution function}
}
\description{
vector of values that follow specified distribution
}
\examples{
distribution_vector(10, "normal", list(mean = 2, sd = 0.5))
}
