% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{special_vector}
\alias{special_vector}
\title{Wrapper that allows generating a special type vectors}
\usage{
special_vector(size, type, configuration)
}
\arguments{
\item{size}{integer, vector length}

\item{type}{type of vector, one of: "id", "distribution"}

\item{configuration}{list of arguments required by vector function}
}
\description{
Wrapper that allows generating a special type vectors
}
\examples{
special_vector(10, "id", list(start = 3))
}
